requirements.txt
blinker==1.8.2
chardet==5.2.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
Flask-WTF==1.2.1
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
PyPDF2==3.0.1
python-dotenv==1.0.1
reportlab==4.2.2
Werkzeug==3.0.3
WTForms==3.1.2
